requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
httpx==0.28.1
python-dotenv==1.0.1
pandas==2.2.0
numpy==1.26.3
google-generativeai==0.8.4
pydantic==2.10.6
yfinance==0.2.55